api/requirements.txt
Django==4.1.5
djangorestframework==3.14.0
PyMySQL==1.0.2
django-cors-headers
